Moved get_arrival_chart() to seperate script
